# Improve StringToUTF16 performance by ~2-3x